fix: Facade 패턴으로 트랜잭션 내부 S3/Redis 호출 분리
