+++ src/main/java/io/wisoft/ignoa_api/user/controller/UserController.java
@@ -3,7 +3,9 @@ package io.wisoft.ignoa_api.user.controller;
 import io.wisoft.ignoa_api.global.common.ApiResponse;
 import io.wisoft.ignoa_api.user.dto.request.UpdateUserRequest;
 import io.wisoft.ignoa_api.user.dto.response.UserMeResponse;
-import io.wisoft.ignoa_api.user.service.UserService;
+import io.wisoft.ignoa_api.user.service.UserCommandService;
+import io.wisoft.ignoa_api.user.service.UserFacade;
+import io.wisoft.ignoa_api.user.service.UserQueryService;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -22,25 +24,27 @@ import static io.wisoft.ignoa_api.global.common.CookieUtils.createClearRefreshTo
 @RequiredArgsConstructor
 public class UserController {
 
-    private final UserService userService;
+    private final UserCommandService userCommandService;
+    private final UserQueryService userQueryService;
+    private final UserFacade userFacade;
 
     @GetMapping("/email/duplicate")
     public ResponseEntity<ApiResponse<Void>> checkDuplicateEmail(@RequestParam String email) {
-        userService.checkDuplicateEmail(email);
+        userQueryService.checkDuplicateEmail(email);
         ApiResponse<Void> response = ApiResponse.of(null, "사용 가능한 이메일입니다.");
         return ResponseEntity.ok(response);
     }
 
     @GetMapping("/nickname/duplicate")
     public ResponseEntity<ApiResponse<Void>> checkDuplicateNickname(@RequestParam String nickname) {
-        userService.checkDuplicateNickname(nickname);
+        userQueryService.checkDuplicateNickname(nickname);
         ApiResponse<Void> response = ApiResponse.of(null, "사용 가능한 닉네임입니다.");
         return ResponseEntity.ok(response);
     }
 
     @GetMapping("/me")
     public ResponseEntity<ApiResponse<UserMeResponse>> getMe(@AuthenticationPrincipal Long userId) {
-        UserMeResponse data = userService.getMe(userId);
+        UserMeResponse data = userQueryService.getMe(userId);
         ApiResponse<UserMeResponse> response = ApiResponse.of(data, "마이페이지 조회에 성공했습니다.");
         return ResponseEntity.ok(response);
     }
@@ -50,14 +54,14 @@ public class UserController {
             @AuthenticationPrincipal Long userId,
             @RequestPart MultipartFile image
     ) {
-        UserMeResponse data = userService.updateProfileImage(userId, image);
+        UserMeResponse data = userFacade.updateProfileImage(userId, image);
         ApiResponse<UserMeResponse> response = ApiResponse.of(data, "프로필 사진이 변경되었습니다.");
         return ResponseEntity.ok(response);
     }
 
     @DeleteMapping("/me/profile-image")
     public ResponseEntity<ApiResponse<Void>> deleteProfileImage(@AuthenticationPrincipal Long userId) {
-        userService.deleteProfileImage(userId);
+        userCommandService.deleteProfileImage(userId);
         ApiResponse<Void> response = ApiResponse.of(null, "프로필 사진이 삭제되었습니다.");
         return ResponseEntity.ok(response);
     }
@@ -67,7 +71,7 @@ public class UserController {
             @AuthenticationPrincipal Long userId,
             @Valid @RequestBody UpdateUserRequest request
     ) {
-        UserMeResponse data = userService.patchMe(userId, request);
+        UserMeResponse data = userCommandService.patchMe(userId, request);
         ApiResponse<UserMeResponse> response = ApiResponse.of(data, "유저 정보가 수정되었습니다.");
         return ResponseEntity.ok(response);
     }
@@ -79,7 +83,7 @@ public class UserController {
             @CookieValue("refresh_token") String refreshToken,
             HttpServletResponse response
     ) {
-        userService.deleteMe(userId, authHeader.substring(7), refreshToken);
+        userFacade.deleteMe(userId, authHeader.substring(7), refreshToken);
 
         ResponseCookie clearCookie = createClearRefreshTokenCookie();
         response.addHeader(HttpHeaders.SET_COOKIE, clearCookie.toString());
