+++ src/main/java/io/wisoft/ignoa_api/user/service/UserFacade.java
@@ -1,0 +1,40 @@
+package io.wisoft.ignoa_api.user.service;
+
+import io.wisoft.ignoa_api.auth.service.RefreshTokenService;
+import io.wisoft.ignoa_api.auth.service.TokenBlacklistService;
+import io.wisoft.ignoa_api.global.infra.storage.StorageService;
+import io.wisoft.ignoa_api.user.dto.response.UserMeResponse;
+import io.wisoft.ignoa_api.user.entity.User;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import org.springframework.web.multipart.MultipartFile;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class UserFacade {
+
+    private final StorageService storageService;
+    private final RefreshTokenService refreshTokenService;
+    private final TokenBlacklistService tokenBlacklistService;
+    private final UserCommandService userCommandService;
+    private final UserQueryService userQueryService;
+
+    public UserMeResponse updateProfileImage(Long userId, MultipartFile image) {
+        User user = userQueryService.findById(userId);
+
+        String oldImageUrl = user.getProfileImageUrl();
+        String newImageUrl = storageService.upload(image);
+        user.updateProfileImage(newImageUrl);
+        userCommandService.saveProfileImage(user, oldImageUrl);
+
+        return UserMeResponse.from(user);
+    }
+
+    public void deleteMe(Long userId, String accessToken, String refreshToken) {
+        userCommandService.withdraw(userId);
+        tokenBlacklistService.blacklist(accessToken);
+        refreshTokenService.delete(refreshToken);
+    }
+}
